+++ app/src/docs/rangeDatePicker/rangeDatePickerExamples.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { uuiDayjs } from '../../helpers/dayJsHelper';
 import type { Dayjs } from '../../helpers/dayJsHelper';
 import { Day, DayProps } from '@epam/uui-components';
-import { cx } from '@epam/uui-core';
+import { cx, RangeDatePickerValue } from '@epam/uui-core';
 import { IPropSamplesCreationContext } from '@epam/uui-docs';
 import {
-    Text, RangeDatePickerProps, RangeDatePickerValue, rangeDatePickerPresets,
+    Text, RangeDatePickerProps, rangeDatePickerPresets,
     uuiRangeDatePickerBody,
 } from '@epam/uui';
 import css from './rangeDatePickerExamples.module.scss';