+++ uui/components/datePickers/RangeDatePickerBody.tsx
@@ -1,7 +1,6 @@
 import React, { forwardRef, useState, type JSX } from 'react';
 import cx from 'classnames';
-import type { IControlled, RangeDatePickerPresets } from '@epam/uui-core';
-import type { DayProps } from '@epam/uui-components';
+import type { IControlled, RangeDatePickerPresets, DayProps, RangeDatePickerInputType, RangeDatePickerValue } from '@epam/uui-core';
 import { uuiDaySelection, Day } from '@epam/uui-components';
 import { FlexCell, FlexRow } from '../layout';
 import { CalendarPresets } from './CalendarPresets';
@@ -12,10 +11,7 @@ import { uuiDayjs } from '../../helpers/dayJsHelper';
 import {
     defaultRangeValue, getMonthOnOpen, getWithFrom, getWithTo, uuiDatePickerBodyBase, valueFormat,
 } from './helpers';
-import type {
-    CommonDatePickerBodyProps,
-    RangeDatePickerInputType, RangeDatePickerValue, RangeDatePickerBodyValue, ViewType,
-} from './types';
+import type { CommonDatePickerBodyProps, ViewType } from './types';
 
 import css from './RangeDatePickerBody.module.scss';
 
@@ -92,6 +88,20 @@ export const rangeDatePickerPresets: RangeDatePickerPresets = {
         }),
     },
 };
+
+/**
+ * Represents date picker body value
+ */
+export interface RangeDatePickerBodyValue<TSelection> {
+    /**
+     * Currently setting date
+     */
+    inFocus: RangeDatePickerInputType;
+    /**
+     * Date currently set
+     */
+    selectedDate: TSelection;
+}
 
 export interface RangeDatePickerBodyProps<T> extends CommonDatePickerBodyProps, IControlled<RangeDatePickerBodyValue<T>> {
     renderFooter?(): React.ReactNode;