+++ uui/components/datePickers/index.ts
@@ -6,7 +6,3 @@ export * from './RangeDatePicker';
 export * from './RangeDatePickerBody';
 
 export type { DatePickerBodyProps } from './DatePickerBody';
-export type {
-    RangeDatePickerValue,
-    RangeDatePickerBodyValue,
-} from './types';