+++ changelog.md
@@ -1,3 +1,11 @@
+# 6.xx.xx - xx.xx.2025
+
+**What's New**
+* [DatePicker]: added `preventEmpty` prop to prevent picker form being empty. See example [here](https://uui.epam.com/documents?id=datePicker&mode=doc&category=components#examples-datePicker-PreventEmpty)
+* [RangeDatePicker]: added `preventEmptyFromDate` and `preventEmptyToDate` props to prevent the RangeDatePicker from having an empty 'from' or 'to' date. See example [here](https://uui.epam.com/documents?id=rangeDatePicker&mode=doc&category=components#examples-rangeDatePicker-PreventEmpty)
+
+
+
 # 6.0.0 - 23.04.2025
 
 **What's New**